requirements.txt
accelerate==1.8.1
acres==0.3.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.0
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
blis==1.2.0
Bottleneck==1.4.2
bs4==0.0.2
cachetools==6.1.0
catalogue==2.0.10
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
ci-info==0.3.0
clean-text==0.6.0
click==8.1.7
cloudpathlib==0.20.0
confection==0.1.5
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.0
convokit==2.5.3
cycler==0.12.1
cymem==2.0.11
dash==3.2.0
dash-bootstrap-components==2.0.4
dataclasses-json==0.6.7
DateTime==5.5
dill==0.3.9
distro==1.9.0
dnspython==2.7.0
emoji==1.7.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
etelemetry==0.3.1
exceptiongroup==1.2.2
faiss-cpu==1.9.0
filelock==3.16.1
Flask==3.1.2
fonttools==4.56.0
frontend==0.0.3
frozenlist==1.4.1
fsspec==2024.9.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
h5py==3.12.1
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.25.2
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.7
langchain==0.3.27
langchain-community>=0.3.9
langsmith>=0.3.45,<1.0.0
langchain-ollama==0.3.10
langcodes==3.5.0
langgraph==0.4.8
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.4.32
language_data==1.3.0
looseversion==1.3.0
lxml==5.4.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.1
marshmallow==3.22.0
matplotlib==3.9.4
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
msgpack-numpy==0.4.8
multidict==6.1.0
murmurhash==1.0.12
mypy-extensions==1.0.0
narwhals==1.44.0
nest-asyncio==1.6.0
networkx==3.2.1
nibabel==5.3.2
nipype==1.10.0
nltk==3.9.1
numexpr==2.10.2
numpy==1.26.4
ollama==0.6.0
openai==1.51.2
openpyxl==3.1.5
orjson==3.10.7
ormsgpack==1.10.0
packaging==24.1
pandas==2.2.3
pathlib==1.0.1
pillow==10.4.0
plotly==6.3.1
preshed==3.0.9
propcache==0.2.0
protobuf==6.31.1
prov==2.0.1
psutil==7.0.0
puremagic==1.28
pyarrow==20.0.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pydeck==0.9.1
pydot==3.0.4
Pygments==2.19.1
pymongo==4.11.1
PyMuPDF==1.25.5
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-nmap==0.7.1
pytz==2024.2
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==6.3.2
referencing==0.36.2
regex==2024.9.11
reportlab==4.4.0
requests==2.32.3
requests-toolbelt==1.0.0
retrying==1.4.2
rich==13.9.4
rpds-py==0.25.1
ruff==0.9.9
safetensors==0.4.5
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==3.1.1
shellingham==1.5.4
simplejson==3.20.1
six==1.16.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.35
srsly==2.5.1
starlette==0.46.2
streamlit==1.46.1
sympy==1.13.3
tenacity==8.5.0
textblob==0.18.0.post0
thinc==8.3.4
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.0
toml==0.10.2
torch==2.4.1
tornado==6.5.1
tqdm==4.66.5
traits==7.0.2
transformers==4.45.2
typer==0.15.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
Unidecode==1.3.8
urllib3==2.2.3
uvicorn==0.34.2
virustotal-api==1.1.11
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.2
xmltodict==0.14.1
xxhash==3.5.0
yarl==1.14.0
zipp==3.21.0
zope.interface==7.1.0
zstandard==0.23.0